Update README.md
The traditional pen and paper game of Tic Tac Toe has been technologically implemented :application. Players mark X's and O's in a 3x3 grid in turn, trying to get three of their markings in a row horizontally, vertically, or diagonally-before their rival. Usually, the game has two player multiplayer mode or single player mode against an Al opponent. The game concludes when one person creates a winning pattern or when the grid is filled in without a winner, leading to a tie. The interface shows a grid where participants can tap or click to add their marks.